refactor!: change namespace
BREAKING CHANGE: changed namespaces from Intermax\LaravelApi\JsonApi to Intermax\LaravelJsonApi
