--- phpstan-baseline.neon
@@ -1,26 +1,0 @@
-parameters:
-	ignoreErrors:
-		-
-			message: "#^Method Intermax\\\\LaravelApi\\\\JsonApi\\\\Filters\\\\FilterResolver\\:\\:resolve\\(\\) has parameter \\$query with generic class Illuminate\\\\Database\\\\Eloquent\\\\Builder but does not specify its types\\: TModelClass$#"
-			count: 1
-			path: src/JsonApi/Filters/FilterResolver.php
-
-		-
-			message: "#^Class Intermax\\\\LaravelApi\\\\JsonApi\\\\Filters\\\\OperatorFilter implements generic interface Spatie\\\\QueryBuilder\\\\Filters\\\\Filter but does not specify its types\\: TModelClass$#"
-			count: 1
-			path: src/JsonApi/Filters/OperatorFilter.php
-
-		-
-			message: "#^Method Intermax\\\\LaravelApi\\\\JsonApi\\\\Filters\\\\OperatorFilter\\:\\:__invoke\\(\\) has parameter \\$query with generic class Illuminate\\\\Database\\\\Eloquent\\\\Builder but does not specify its types\\: TModelClass$#"
-			count: 1
-			path: src/JsonApi/Filters/OperatorFilter.php
-
-		-
-			message: "#^Class Intermax\\\\LaravelApi\\\\JsonApi\\\\Resources\\\\IncludesBag implements generic interface Illuminate\\\\Contracts\\\\Support\\\\Arrayable but does not specify its types\\: TKey, TValue$#"
-			count: 1
-			path: src/JsonApi/Resources/IncludesBag.php
-
-		-
-			message: "#^Property Intermax\\\\LaravelApi\\\\JsonApi\\\\Resources\\\\IncludesBag\\:\\:\\$items with generic class Illuminate\\\\Support\\\\Collection does not specify its types\\: TKey, TValue$#"
-			count: 1
-			path: src/JsonApi/Resources/IncludesBag.php
