+++ phpstan.neon
@@ -1,6 +1,5 @@
 includes:
     - %currentWorkingDirectory%/vendor/nunomaduro/larastan/extension.neon
-    - phpstan-baseline.neon
 
 parameters:
