+++ src/Includes/Relation.php
@@ -1,0 +1,18 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Intermax\LaravelJsonApi\Includes;
+
+class Relation implements Contracts\Relation
+{
+    public function __construct(
+        protected string $name
+    ) {
+    }
+
+    public function allowedInclude(): string
+    {
+        return $this->name;
+    }
+}
