--- src/JsonApi/Exceptions/Contracts/RenderableError.php
@@ -1,16 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Intermax\LaravelApi\JsonApi\Exceptions\Contracts;
-
-use Intermax\LaravelApi\JsonApi\Resources\ErrorResource;
-use Intermax\LaravelApi\JsonApi\Resources\ErrorResourceCollection;
-
-interface RenderableError
-{
-    /**
-     * @return ErrorResourceCollection<ErrorResource>
-     */
-    public function toResource(): ErrorResourceCollection;
-}
