--- src/JsonApi/Includes/Relation.php
@@ -1,18 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Intermax\LaravelApi\JsonApi\Includes;
-
-class Relation implements Contracts\Relation
-{
-    public function __construct(
-        protected string $name
-    ) {
-    }
-
-    public function allowedInclude(): string
-    {
-        return $this->name;
-    }
-}
