--- src/JsonApi/Resources/RelationType.php
@@ -1,16 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Intermax\LaravelApi\JsonApi\Resources;
-
-class RelationType
-{
-    public const MANY = 'many';
-
-    public const ONE = 'one';
-
-    private function __construct()
-    {
-    }
-}
