+++ src/Resources/RelationType.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Intermax\LaravelJsonApi\Resources;
+
+class RelationType
+{
+    public const MANY = 'many';
+
+    public const ONE = 'one';
+
+    private function __construct()
+    {
+    }
+}
